Time: 1 ms (98.23%), Space: 43.1 MB (10.53%) - LeetHub

+++ 1295-find-numbers-with-even-number-of-digits/1295-find-numbers-with-even-number-of-digits.java
@@ -1,6 +1,6 @@
 class Solution {
     public int findNumbers(int[] nums) {
-        short count = 0;
+        byte count = 0;
         for(int i:nums) {
             if((i>9 && i<100) || (i>999 && i<10000) || (i==100000)) count++;
         }
